Split Links controllers
