+++ src/controllers/links/Messages.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace flusio\controllers;
+namespace flusio\controllers\links;
 
 use Minz\Response;
 use flusio\models;
@@ -12,7 +12,7 @@ use flusio\utils;
  * @author  Marien Fressinaud <dev@marienfressinaud.fr>
  * @license http://www.gnu.org/licenses/agpl-3.0.en.html AGPL
  */
-class LinkMessages
+class Messages
 {
     /**
      * @request_param string link_id
